SV-656 separate application summary domain model from API return model so that we can expose Version as a list of strings on the API from the string of Json that Dapper returns from the database.
