+++ src/SFA.DAS.AssessorService.Application/Handlers/Apply/Review/OrganisationApplicationsHandler.cs
@@ -1,9 +1,12 @@
-﻿using MediatR;
+﻿using AutoMapper;
+using MediatR;
 using Microsoft.Extensions.Logging;
 using SFA.DAS.AssessorService.Api.Types.Models.Apply.Review;
 using SFA.DAS.AssessorService.Application.Interfaces;
 using SFA.DAS.AssessorService.ApplyTypes;
+using SFA.DAS.AssessorService.Domain.Entities;
 using SFA.DAS.AssessorService.Domain.Paging;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -28,7 +31,9 @@ namespace SFA.DAS.AssessorService.Application.Handlers.Apply.Review
             var organisationApplicationsResult = await _repository.GetOrganisationApplications(request.ReviewStatus, request.SortColumn, request.SortAscending,
                 request.PageSize, request.PageIndex);
 
-            return new PaginatedList<ApplicationSummaryItem>(organisationApplicationsResult.PageOfResults.ToList(),
+            var items = Mapper.Map<IEnumerable<ApplicationListItem>, IEnumerable<ApplicationSummaryItem>>(organisationApplicationsResult.PageOfResults);
+
+            return new PaginatedList<ApplicationSummaryItem>(items.ToList(),
                     organisationApplicationsResult.TotalCount, request.PageIndex, request.PageSize, request.PageSetSize);
         }
     }
