+++ src/SFA.DAS.AssessorService.Domain/Entities/ApplicationListItem.cs
@@ -1,9 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
 
-namespace SFA.DAS.AssessorService.ApplyTypes
+namespace SFA.DAS.AssessorService.Domain.Entities
 {
-    public class ApplicationSummaryItem
+    // Represents a row in the Apply_List_Applications stored proc response
+    public class ApplicationListItem
     {
         public Guid ApplicationId { get; set; }
         public int SequenceNo { get; set; }
@@ -11,17 +11,16 @@ namespace SFA.DAS.AssessorService.ApplyTypes
         public string StandardName { get; set; }
         public int? StandardCode { get; set; }
         public string StandardReference { get; set; }
-        public string Standard => StandardCode.HasValue ? $"{StandardName} ({StandardCode})" : StandardName;
-        public List<string> Versions { get; set; }
+        public string Versions { get; set; }
         public DateTime? SubmittedDate { get; set; }
         public DateTime? FeedbackAddedDate { get; set; }
         public DateTime? ClosedDate { get; set; }
         public int SubmissionCount { get; set; }
         public string ApplicationStatus { get; set; }
+        public string ApplicationType { get; set; }
         public string ReviewStatus { get; set; }
         public string FinancialStatus { get; set; }
         public string FinancialGrade { get; set; }
-        public string SequenceStatus { get; set; } // NOTE: Only used for Closed Applications
-        public string ApplicationType { get; set; }
+        public string SequenceStatus { get; set; }
     }
 }
